ORM abstraction layer superfluous as I'm specializing this gem for AR
